feat: Add Redirect component and refactor App component
This commit introduces a new Redirect component that handles the redirection logic based on the short URL provided. The Redirect component makes API calls to retrieve the long URL and update metadata accordingly. 

The changes also include refactoring the App component to remove the previous redirection logic and utilize the new Redirect component. This refactor improves code organization and separates concerns for better maintainability.
